ml-service/requirements.txt
fastapi
uvicorn
motor
pymongo
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
joblib
python-dotenv
